backend/pyproject.toml
[tool.black]
line-length = 78

[flake8]
max-line-length = 78


[tool.poetry]
name = "Chickie"
version = "0.1.0"
description = "Aplicação para pedidos e entregas de refeições"
authors = ["Antonio Fernando <antoniofernandodj@outlook.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "3.10.12"
sqlalchemy = "^2.0.19"
aiopg = "^1.4.0"
fastapi = "^0.100.1"
dynaconf = "^3.2.0"
uvicorn = "^0.23.1"
cryptography = "^41.0.2"
redis = "^4.6.0"
bcrypt = "^4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
pika = "^1.3.2"
httpx = "^0.24.1"
weasyprint = "^59.0"
jinja2 = "^3.1.2"
pytz = "^2023.3"
aioamqp = "^0.15.0"
pymongo = "^4.4.1"
celery = "^5.3.1"
dropbox = "^11.36.2"


[tool.poetry.group.dev.dependencies]
faker = "^19.2.0"
mypy = "^1.4.1"
flake8 = "^6.0.0"
pytest = "^7.4.0"
types-redis = "^4.6.0.3"
types-passlib = "^1.7.7.12"
types-python-jose = "^3.3.4.8"
pyside6 = "^6.5.2"
trio = "^0.22.2"
pytest-asyncio = "^0.21.1"
pytest-trio = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
